only clean report folder in writing mode

+++ geb/model.py
@@ -121,8 +121,9 @@ class GEBModel(HazardDriver, ABM, Hydrology):
         self.report_folder = (
             Path(self.config["general"]["report_folder"]) / self.run_name
         )
-        # clean report model at start of run
-        shutil.rmtree(self.report_folder, ignore_errors=True)
+        if self.mode == "w":
+            # clean report model at start of run
+            shutil.rmtree(self.report_folder, ignore_errors=True)
         self.report_folder.mkdir(parents=True, exist_ok=True)
 
         self.spinup_start = datetime.datetime.combine(
